+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ What's new
 -----
 
 * ``Crystal.ideal`` and ``Crystal.primitive`` will now preserve subclasses.
+* ``Crystal.primitive`` will now always return a new ``Crystal``, even if the ``Crystal`` already has a primitive cell.
 
 0.6.3
 -----